[dev]: Rewrite logic of connect button
